index system

+++ lib/lib-table-cluster/src/main/java/org/sagebionetworks/table/cluster/utils/ColumnConstants.java
@@ -79,5 +79,5 @@ public class ColumnConstants {
 	
 	public static final String CHARACTER_SET_UTF8_COLLATE_UTF8_GENERAL_CI = "CHARACTER SET utf8 COLLATE utf8_general_ci";
 	
-	public static final long MAX_MYSQL_VARCHAR_INDEX_LENGTH = 255;
+	public static final int MAX_MYSQL_VARCHAR_INDEX_LENGTH = 255;
 }
